refactor: auth -> user
