+++ lib/app/modules/user/data/repositories/flutter_secure_storage_token_repository.dart
@@ -3,7 +3,7 @@ import 'dart:async';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:injectable/injectable.dart';
 import 'package:rxdart/rxdart.dart';
-import 'package:ziggle/app/modules/auth/domain/repositories/token_repository.dart';
+import 'package:ziggle/app/modules/user/domain/repositories/token_repository.dart';
 
 @Singleton(
   as: TokenRepository,
